Simplify pattern of characters to quote to avoid non-ascii characters, correcting an error when used under ruby 1.9.
Adding the modeline caused ruby 1.9 to being interpreting the source
file as utf-8, which then caused the regex, which contained invalid
utf-8 sequences, to become invalid.

This fix inverts the character class in the regex, allowing us to
specify pure ascii characters which should be unaffected by the file
encoding. As a benefit, the character class is shorter and easier to
interpret.