Added reply_to_author option in the example config.
Added a reply_to_author setting in the git-notifier-config.example.yml

Setting this field to true enables generation of Reply-To: header in the
email with the alias and address same as the author of the commit.

If this field is instead false or unset, the Reply-To: header is
generated with the same information as the From: header.