Add the Reply-To header in the email.
Reply-To header is added to the email based on the reply_to_address
instance variable.

If either reply_to_author config is set to false/unset, OR from_alias is
NIL, Reply-To: header is filled with just the reply_to_address.

Otherwise, Reply-To: header is filled in the
from_alias <reply_to_address> format.